requirements.txt
flask>=3.0
python-dotenv>=1.0
aiohttp>=3.9
solders>=0.21
solana>=0.34
